requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.5
opencv-python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
scikit-learn==1.0.2
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
Werkzeug==2.2.2
